fetch link chnaged

+++ src/Routes/ManageAllOrder/ManageAllOrder.js
@@ -10,26 +10,28 @@ const ManageAllOrder = ({ allorder }) => {
 
     // destructuring 
     const { completed, _id, title, } = allorder;
-   
+
 
     // Delete a place
     const handledeletplace = (id) => {
-        const uri = `https://floating-dawn-88339.herokuapp.com/service/${id}`;
-        fetch(uri, {
-            method: 'DELETE'
-        })
-            .then(res=>res.json())
-            .then(data=>{
-                if(data.deletedCount>0){
-                    alert('Service Delete Successfully')
-                    window.location.reload()
-                   
-                }
+        const decision = window.confirm("Are you sure want to delete ?")
+        if (decision) {
+            const uri = `https://tourdaw-server.herokuapp.com/services/${id}`;
+            fetch(uri, {
+                method: 'DELETE'
             })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount > 0) {
+                        alert('Service Delete Successfully')
+                        window.location.reload()
 
+                    }
+                })
+        }
     }
-    
-    
+
+
 
     return (
 
@@ -38,11 +40,11 @@ const ManageAllOrder = ({ allorder }) => {
             className="d-flex justify-content-between align-items-start">
             <div className="ms-2 me-auto">
                 <div className="fw-bold">{title}</div>
-                <b>Status:<strong>{}</strong></b>
+                <b>Status:<strong>{ }</strong></b>
             </div>
             <span>
                 <Button onClick={() => handledeletplace(_id)} className="btn btn-danger rounded">{trash}</Button>
-                <Button  className="btn btn-primary rounded ms-3">{check}</Button>
+                <Button className="btn btn-primary rounded ms-3">{check}</Button>
             </span>
         </ListGroup.Item>
     );
